move config to grafana runtime
