+++ packages/grafana-runtime/src/index.ts
@@ -1,1 +1,3 @@
 export * from './services';
+export { config, Settings } from './config';
+export { loadPluginCss } from './utils/plugin';
